When explorer opens a folder, it tries to read desktop.ini from every sub-folder, fetching many zaps scattered on the disk.
First time it is still going to be slow, but now the data stays cached at least.

+++ trunk/zfs-win/DataSet.h
@@ -31,7 +31,7 @@ namespace ZFS
 		Pool* m_pool;
 		std::map<std::wstring, dnode_phys_t> m_cache;
 
-		bool Init(ObjectSet& os, const char* name = NULL, size_t root_index = -1);
+		bool Init(ObjectSet& os, const char* name = NULL, uint64_t root_index = -1);
 		void RemoveAll();
 		void SetDefaults(DataSet* parent);
